% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{agglomerative}
\alias{agglomerative}
\title{Apply agglomerative clustering to time series}
\usage{
agglomerative(dm, k)
}
\arguments{
\item{dm}{is a matrix or distance matrix object}

\item{k}{number of clusters}
}
\description{
`agglomerative` apply complete linkage agglomerative clustering algorithm based on a distance matrix
}
\details{
this function uses `hclust` and `cutree` from kmed package. The complete linkage is used.
}
