% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{compute_dm}
\alias{compute_dm}
\title{Computes distances between time-series}
\usage{
compute_dm(M_pred, M_se = NULL, DTW = F, type = "l1")
}
\arguments{
\item{M_pred}{a data.frame of time-series, where each column represents a time-series}

\item{M_se}{a data.frame of standard errors representing the level of confidence for each interpolated data points on M_pred. use if standard error is calculated for M_pred.}

\item{DTW}{whether to apply dynamic time-warping or not}

\item{type}{either "l1" vs "hellinger", l1 distance or hellinger distance is applied}
}
\description{
`compute_dm` calculates pairwise distance between time series and returns a distance matrix.
}
