% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{gp_interpolation}
\alias{gp_interpolation}
\title{Gaussian Process regression with Matern kernel is used to interpolate time series}
\usage{
gp_interpolation(dataframe, n_intervals)
}
\arguments{
\item{dataframe}{a data.frame with each column representing time-series with missing value as NA.}

\item{n_intervals}{the number of intervals to be interpolated. This is used to rescale your time series on a regular interval.}
}
\description{
`gp_interpolation` returns a data.frame with interpolated data points based on Gaussian Process regression.
}
\details{
`GauPro_kernel_model` and `Matern52` function from the GauPro package was used to inteproalte data points.
}
