% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{hellinger}
\alias{hellinger}
\title{squared Hellinger distances between two normal distributions}
\usage{
hellinger(mu1, se1, mu2, se2)
}
\arguments{
\item{mu1}{mean of the first normal distribution}

\item{se1}{standard error of the first normal distribution}

\item{mu2}{mean of the second normal distribution}

\item{se2}{standard error of the second normal distribution}
}
\description{
`hellinger` computes and returns squared Hellinger distance between two normal distributions
}
\details{
The squared Hellinger distance between two normal distributions, \eqn{X \sim N(\mu_x,\sigma_x^2)} and \eqn{Y \sim N(\mu_y,\sigma_y^2)} \deqn{H^2(X,Y) = 1-\sqrt{\frac{2\sigma_x\sigma_y}{\sigma_x^2+\sigma_2^2}}e^{\frac{-(\mu_x -\mu_y)^2}{4*(\sigma_x^2 + \sigma_y^2)}}}
}
