% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{k_medoid}
\alias{k_medoid}
\title{Apply k-medoid clustering to time series}
\usage{
k_medoid(dm, k)
}
\arguments{
\item{dm}{is a matrix or distance matrix object}

\item{k}{number of clusters}
}
\description{
`kmedoid` apply k-medoid clustering algorithm based on a distance matrix
}
\details{
this function uses `fastkmed` from kmed package.
}
