% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{kmeans_ts}
\alias{kmeans_ts}
\title{Apply k-means clustering to time series}
\usage{
kmeans_ts(df_timeseries, k, DTW = F)
}
\arguments{
\item{df_timeseries}{a data.frame where each column represents time series.}

\item{k}{number of clusters}

\item{DTW}{whether to apply dynamic time-warping or not}
}
\description{
`kmeans` apply k-means clustering algorithm to a collection of time series and return a vector representing clusters
}
\details{
when DTW is not applied `kmeans` function from the stat package is used. when DTW is applied, `dtwclust` function from tsclust package is used.
}
