% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{linearly interpolate time series with missing value}
\usage{
linear_interpolation(dataframe, n_intervals)
}
\arguments{
\item{dataframe}{a data.frame with each column representing time-series with missing value as NA. The first and the last item must exist.}

\item{n_intervals}{the number of intervals to be interpolated. This is used to rescale your time series on a regular interval.}
}
\description{
`linear_interpolation` is used to linearly interpolate time series with missing value.
}
\details{
`approx` function from the stats package was used to linearly inteproalte data points.
}
